Refactor map render logic into standalone functions
