+++ src/meshapi/urls.py
@@ -33,9 +33,9 @@ urlpatterns = [
     path("query/buildings/", views.QueryBuilding.as_view(), name="meshapi-v1-query-building"),
     path("query/members/", views.QueryMember.as_view(), name="meshapi-v1-query-member"),
     path("query/installs/", views.QueryInstall.as_view(), name="meshapi-v1-query-install"),
-    path("mapdata/nodes/", views.MapDataNodeList.as_view(), name="meshapi-v1-map-data-installs"),
-    path("mapdata/links/", views.MapDataLinkList.as_view(), name="meshapi-v1-map-data-links"),
-    path("mapdata/sectors/", views.MapDataSectorList.as_view(), name="meshapi-v1-map-data-sectors"),
+    path("mapdata/nodes/", views.map_data_node_list, name="meshapi-v1-map-data-installs"),
+    path("mapdata/links/", views.map_data_link_list, name="meshapi-v1-map-data-links"),
+    path("mapdata/sectors/", views.map_data_sector_list, name="meshapi-v1-map-data-sectors"),
     path("geography/whole-mesh.kml", views.WholeMeshKML.as_view(), name="meshapi-v1-geography-whole-mesh-kml"),
     path("update-panoramas/", views.update_panoramas, name="meshapi-v1-update-panoramas"),
 ]
